Fix float values provided in json silently converting into ints

+++ src/magic_enchantment.h
@@ -375,7 +375,7 @@ class enchant_cache : public enchantment
                                     const std::map<TKey, double> &mult_map ) const;
 
         void serialize( JsonOut &jsout ) const;
-        void add_value_add( enchant_vals::mod value, int add_value );
+        void add_value_add( enchant_vals::mod value, float add_value );
 
         void set_has( enchantment::has value );
         void add_value_mult( enchant_vals::mod value, float mult_value );
